api/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
torch>=2.0.0
nltk==3.8.1
locust==2.17.0
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.25.0 